Move MemoryFileWrapper into its own file.
